Snapshot of develop 2019-08-02 (Internal repo commit 02e8e5a)
* 2019-08-02 (02e8e5a) More sanitization tweaks
* 2019-08-02 (1c8d9ae) Tweak hostname flaw
*   2019-07-31 (0174053) Merge branch 'rpmlint' into develop
|\
| * 2019-07-31 (d518c0a) Cleaned up top-level Makefile; added driver RPM instructions
| * 2019-07-31 (38298bd) Some rpmlint touch-ups
* | 2019-07-30 (5ad0036) Sanitization: Add tarball to dumping and fix merge issues (NoJenkins)
* | 2019-07-30 (f4d09d2) AV-2500: Misc clean-up noted when merging
* |   2019-07-30 (2451d50) Merge branch 'feature--AV-2500-add-sha256sums-to-metadata' into develop
|\ \
| |/
|/|
| * 2019-07-29 (a5e4703) Revert "AV-2500 (but not) Attempt to fix the "sorenb" problems C6 is hitting"
| * 2019-07-29 (f3c27c2) Revert "AV-2500: Don't do prereqs either"
| * 2019-07-29 (5af4b57) Revert "AV-2500: Fix prereq filtering in RPM debug generation"
| * 2019-07-29 (9973433) AV-2500: Document "ocpixml check" minimally and use standard returns
| * 2019-07-29 (fa047a1) AV-2500: Fix prereq filtering in RPM debug generation
| * 2019-07-29 (757a21c) AV-2500: Don't do prereqs either
| * 2019-07-29 (fb89718) AV-2500 (but not) Attempt to fix the "sorenb" problems C6 is hitting
| * 2019-07-29 (9488b72) Cleaner errors on missing sudo / yum
| *   2019-07-29 (8cdd2eb) Merge remote-tracking branch 'origin/develop' into feature--AV-2500-add-sha256sums-to-metadata
| |\
| |/
|/|
* | 2019-07-29 (d047f25) Remove incorrect note
 /
* 2019-07-19 (4f29b40) AV-2500: Hopefully fix CentOS 6 problem
* 2019-07-19 (5caf5b1) AV-2500: Use our own ocpixml command in simulator loading
* 2019-07-19 (f102a2a) AV-2500: New artifact_getPayload() feature added
* 2019-07-19 (58b0edd) AV-2500: WIP for making ocpixml allow extraction of PAYLOAD to stdout
* 2019-07-19 (db7ecc4) OS::FileSystem - fix weird output and made it do more error checks
* 2019-07-19 (2ce973b) AV-2500: Reduce output
* 2019-07-18 (903c0bf) AV-2500: More cleanup; use OCPI exception types more
* 2019-07-18 (c70b3fa) AV-2500: Misc touch-ups
* 2019-07-18 (73ddf95) Remove extra debug info
* 2019-07-18 (a22c9b7) AV-2503: Add metadata checksum to runtime
* 2019-07-17 (17096a4) AV-2500: Fix problems with "extra" info
* 2019-07-17 (a9b51a9) AV-2500: Added extra info and moved to OCPI-specific exceptions
* 2019-07-16 (b8eea82) AV-2500: Templated sha256 wrappers
* 2019-07-16 (8663426) AV-2500 Misc cleanups
* 2019-07-16 (3d5e829) AV-2500: Ensure always 256-bit response
* 2019-07-16 (deb6be5) AV-2500: Fix printing of off_t type on Xilinx platforms
* 2019-07-16 (fe4127f) AV-2500: Stop double-wrapping of artifact
* 2019-07-15 (239424b) Squashed commit of private scratch space for XML metadata AV-2500

+++ runtime/util/ezxml/src/ocpixml_main.cxx
@@ -20,15 +20,20 @@
 
 // Replaces the original ocpixml script with C++ version (AV-2501)
 
+#include <unistd.h>
 #include <cstring> // strcmp
 #include <string>
+#include <memory>
 #include <iostream>
 #include <fstream>
 #include <sstream>
 #include <map>
 #include <stdexcept>
+
+
 #include "OcpiXmlEmbedded.h"
 #include "OcpiOsFileSystem.h" // This is NOT fully OCPI::OS-enabled, e.g. path names, etc.
+#include "OcpiUtilException.h" // OU::Error
 
 namespace OX = OCPI::Util::EzXml;
 namespace OS = OCPI::OS;
@@ -39,14 +44,15 @@ namespace OS = OCPI::OS;
 "Commands are: \n"\
 "  get   - extract the XML from the file to standard output.\n"\
 "  strip - remove the XML from the file while copying artifact to <file2>.\n"\
-"  add   - add the XML file <file2> to the file, in place. file2 can be '-' for standard input.\n"
+"  add   - add the XML file <file2> to the file, in place.\n"\
+"  check - silently verify artifact's XML.\n"\
+"\n<file2> can be '-' to indicate standard input (add) or output (strip)\n"
 
 #define OCPI_OPTIONS \
     CMD_OPTION(log_level,  l, ULong,  0, "<log-level>\n" \
 	                               "set log level during execution, overriding OCPI_LOG_LEVEL")
 
 #include "CmdOption.h"
-
 
 static int mymain(const char **argv) {
   // C++11: const std::map<std::string, int> valid_verbs = {{"get",3}, {"strip",4}, {"add",4}};
@@ -89,28 +95,31 @@ static int mymain(const char **argv) {
 
   // get
   if (0 == strcmp("get", argv[0])) {
-    OX::artifact_getXML(argv[1], xml);
-    if (xml.empty()) {
+    if (!OX::artifact_getXML(argv[1], xml)) {
       std::cerr << "This file \"" << argv[1] << "\" is not an artifact file.\n";
-      return 1;
+      return EXIT_FAILURE;
     }
     std::cout << xml;
   }
 
   // check
   if (0 == strcmp("check", argv[0])) {
-    OX::artifact_getXML(argv[1], xml);
-    return xml.empty() ? 1 : 0;
+    return OX::artifact_getXML(argv[1], xml)?EXIT_SUCCESS:EXIT_FAILURE;
   }
 
   // strip
   if (0 == strcmp("strip", argv[0])) {
-    OS::FileSystem::copy(argv[1], argv[2]);
-    if (!OX::artifact_stripXML(argv[2])) {
-      std::cerr << "This file \"" << argv[1] << "\" is not an artifact file.\n";
-      OS::FileSystem::remove(argv[2]);
-      return 1;
+    if (0 == strcmp("-", argv[2])) { // Special case of stdout
+      if (isatty(fileno(stdout))) throw OCPI::Util::Error("Asked to write binary to a terminal!");
+      OX::artifact_getPayload(argv[1]);
+    } else {
+      OS::FileSystem::copy(argv[1], argv[2]);
+      if (!OX::artifact_stripXML(argv[2])) {
+        std::cerr << "This file \"" << argv[1] << "\" is not an artifact file.\n";
+        OS::FileSystem::remove(argv[2]);
+        return EXIT_FAILURE;
+      }
     }
   }
-  return 0;
+  return EXIT_SUCCESS;
 }
